+++ lib/cute_print/term_width/static.rb
@@ -1,0 +1,21 @@
+module CutePrint
+  class TermWidth
+
+    # A terminal width with a fixed value
+    # @api private
+    class Static
+
+      attr_reader :width
+
+      def initialize(width)
+        @width = width
+      end
+
+      def visible
+        @width
+      end
+
+    end
+
+  end
+end